Added second, fifth and seventh homework folders.
The task of the second homework required me to synthesize a trigger based of the step and a maximum number.

The task of the fifth homework required me to create a program in a CPU simulator based on a personal variant.

The task of the seventh homework required me to estimate the performance of parallel computing based on a personal variant.